Add global dotenv Config & Impl search endpoint & ratelimiter
added global dotenv configuration to make it easy to use env variables
in any file within ll directories. Also, implemented the search endpoint
to be able to search any thing, be it a character, series or movies
the result contain a collection of all types of data that matches the query.
Also added a ratelimiter middleware to not make anyone abuse the server since its public
for any one to interact with through any http agent