fix: adds an express setting to reveal real ip addresses
this has to be doned to fix ratelimiting issues because this
express app runs behind a proxy and without this setting it will always see one ip addresse
which is the proxy's. so instead of actually distinctively ratelimiting each users; it rate limits
all users when there are 15 requests are sent to the ai endpoint in one day.

this change fixes it.